Better method for scanning strings/comments.
Instead of using syntactic-keywords, we scan the entire buffer once through at the beginning with `gnuplot-scan-after-change' and then put it in the `after-change-functions' hook to repeat when necessary.

I can't get this to work in XEmacs because it doesn't have Emacs' "generic string delimiter" and "generic comment delimiter" syntax classes.