Numerous grammar tweaks. Implement `many1' repetition (regex `+').
`many1' is only used in one place, the rule for `linestyle-spec'. This
rule used to be written (with `maybe' forms) in a way that matched the
empty string, causing an infinite loop if it appeared inside of
a (many (either ...))  form would cause an infinite loop.

Using `many1' instead might mean that I can eliminate the check for
infinite loops, whose effects aren't always intuitive.

`gnuplot-test-parse-all-demos' has a success rate of about 86% with
these tweaks.