Cleaned up the tokenizer and got rid of gnuplot-token-at-point.
The tokenizer now returns a list of tokens w/o a separate
`end-of-tokens' marker, so we can just check for (null? tokens) to see
if we've reached the end of the portion of the line being matched.

The matching machine now returns `nil' for failure, and a one-element
list of the list of remaining tokens for success (so '(nil) means that
it succeeded in matching the whole list of tokens). Fixed the
automated tests to work with this new return convention.