More consistent keys between gnuplot.el and gnuplot-context.el
C-c C-d (instead of C-c C-h or C-c C-/) for the info-lookup function
(gnuplot-info-lookup-symbol if context is not loaded,
gnuplot-info-at-point when context is loaded)

gnuplot-info-at-point now prompts for a symbol to look up if called
with a prefix arg or if parsing fails.

C-c C-s (instead of C-c h) for "extended eldoc" in context
mode (gnuplot-help-function)

TAB for completion in comint buffer. (Is this a problem?)

Setup eldoc thru a named defun rather than a lambda in
gnuplot-mode-hook.