Add option to send plot output to a dedicated buffer
The new option is provided by the `gnuplot-inline-image-mode` variable,
which can now be `nil`, `inline`, or `dedicated`.  There are also
corresponding menu options and commands

This variable should probably be renamed using `make-obsolete-variable`,
since it no longer pertains only to images shown inline in the process
buffer.

Also clean up handling of buffers and processes, particularly in
`gnuplot-make-gnuplot-buffer`:

- Use `unless` instead of `or`

- Check for process status of `gnuplot-process` being `run` instead of
  `(get-process gnuplot-process)`, which is a no-op AFAICT

- Use `buffer-live-p` instead of `buffer-name`

- Workaround `set-process-query-on-exit-flag` vs. obsolete
  `process-kill-without-query` functions

- Look up the running process by its buffer, not by name

Getting the process by name when setting up a new comint buffer is
unreliable when the old process has not yet been deleted, because Emacs
will append `<n>` to the name of the new process to make it unique.
Getting the process from its associated buffer does not have this
problem.