Remove version-related variables.
Since we updated the minimum gnuplot program version requirements, a
lot of version-related variables do not make sense anymore since we
are sure to support the "latest" version (from around 2012 that is).
This simplifies the code quite a bit since we don't even need to keep
track of the underlying gnuplot version used (we might need to re-add
it if we start tracking differences between different minor versions
of gnuplot 5, but I doubt it). This commit:

- Removes the gnuplot-program-{,minor-,major-}version variables
- Removes the gnuplot-three-eight-p variable
- Assumes a modern Emacs is used (>= 24.3)
- Removes the gnuplot-show-{,gnuplot-}version functions
- Removes the gnuplot-fetch-version-number (which was broken anyways,
  and prompted the user that gnuplot version 3.7 was assumed
  everytime).
- Removes the gnuplot-setup-menu-and-toolbar function, which is
  useless since we removed the toolbar support.