Do not add lines to kill-ring when trimming the gnuplot buffer (#72)
I noticed that when the number of lines in the gnuplot comint buffer
exceeds the maximum amount of lines, the extra lines are killed using
kill-line, which causes the kill-ring to be polluted with those lines.
This commit replaces kill-ring by delete-region, which does not have
this side effect. The while loop also has been replaced by dotimes for
clarity.