Fix completion-at-point functions
According to the Emacs documentation the completion functions should not
prefilter the candidates themselves. It is up to the completion UI to perform
the filtering. By deferring the filtering to the completion UI various
completion styles like substring, flex or orderless can be used. In contrast,
the old implementation enforced prefix matching.